feat(Login): modify login flow

+++ app/src/main/java/com/zhengdianfang/dazhongbao/models/api/ProductApi.kt
@@ -21,7 +21,7 @@ interface ProductApi {
     @POST("products/push")
     fun pushProduct(@Field("token") token: String, @Field("sharesCode") sharesCodes: String,
                     @Field("companyName") companyName: String, @Field("basicUnitPrice") basicUnitPrice: Double,
-                    @Field("soldCount") soldCount: Int,@Field("limitTime") limitTime: Long,
+                    @Field("soldCount") soldCount: Long,@Field("limitTime") limitTime: Long,
                     @Field("notes") notes: String): Observable<JsonNode>
 
     @FormUrlEncoded
